// frontend/Types/dataTypes.ts
class Stats {
  XP: number = 0;
  GP: number = 0;
  rank: number = 0;
  numGames: number = 0;
  gamesWon: number = 0;
}

export class rankObj {
  rank: string = allRanks[4];
  tier: any = rank4;
}

export class UserTypeNew {
  id: number = 0;
  login: string = "";
  avatar: string = "";
  fullname: string = "";
  stats: Stats = new Stats();
  rank: number = 0;
  GP: number = 0;
  numGames: number = 0;
  gamesWon: number = 0;
  achievement: number[] = [];
  status: string = "";
  relation: string = "";
  constructor() {}
}
// ...
export interface matchDataType {
  login: string;
  avatar: string;
  badge: number;
  fullName: string;
  games: number;
  Win: number;
  GP: number;
  relation: string;
}